Tweak conditions for when the circuit breaker trips
Previously it would only trip if 60% of all requests since the last time
the circuit breaker tripped had failed. We send a lot of requests to
vSphere, and failures don't happen very frequently, so say for example
that some time has passed since the last failure and we've sent 2000
requests to vSphere. If something happens causing all requests to fail,
we still need to send another 3000 requests before the circuit breaker
trips. This commit adds an interval to the circuit breaker, meaning the
numbers get reset every 10 minutes (as long as the circuit breaker is
closed).