Upgrade to Go 1.6.2, remove gometalinter (#14)
* Upgrade to Go 1.6.2

* Remove GO15VENDOREXPERIMENT from .travis.yml

Go 1.6 no longer uses this

* Remove gometalinter from .travis.yml for now

It has too many false duplicates (especially when checking for duplicate
code, and deadlines being exceeded). I think it's a good tool to run
locally before committing to see potential issues, but I'm not sure if
having the build fail because of it is a good idea.