Don't release semaphore until underlying vSphere task has finished
This adds pushback to the API if/when vSphere starts to slow down.

This is done by "forking" off a goroutine when we have created the
initial task to finish, and using a different Context for this to
prevent the task.Wait from returning early if the incoming HTTP
connection times out (which would cancel the context).